pyproject.toml
[project]
name = "free-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-inference>=1.0.0b9",
    "azure-core>=1.35.0",
    "dotenv>=0.9.9",
    "mcp>=1.12.4",
]
